Update UI lifecycle refresh for all components in Dashboard One

+++ database/DeployDatabase/ParkerDB/dbo/Stored Procedures/spLocal_EY_DxH_Insert_ProductionData.sql
@@ -76,8 +76,8 @@ AS
 		@Remaining_BreakMinutes FLOAT,
 		@Remaining_Minutes FLOAT,
 		@TotalRemaining_Minutes FLOAT,
-		@Ideal FLOAT,
-		@Target FLOAT,
+		@Ideal INT,
+		@Target INT,
 		@Order_Start_Time DATETIME,
 		@Previous_Production_Id INT,
 		@Minutes_Less INT;
